requirement.txt
aliyun-python-sdk-core==2.13.4
aliyun-python-sdk-core-v3==2.13.3
aliyun-python-sdk-ecs==4.16.10
aliyun-python-sdk-kms==2.6.0
amqp==2.4.2
appdirs==1.4.4
beautifulsoup4==4.8.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
crcmod==1.7
distlib==0.3.1
Django==2.2.1
django-cors-headers==2.4.0
django-redis==4.10.0
django-redis-sessions==0.6.1
django-rest-swagger==2.2.0
djangorestframework==3.9.4
et-xmlfile==1.0.1
filelock==3.0.12
filetype==1.0.7
Flask==1.1.2
Flask-Cors==3.0.9
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
importlib-metadata==2.0.0
itsdangerous==1.1.0
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
jmespath==0.9.4
jpush==3.3.8
kombu==4.5.0
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
openapi-codec==1.3.2
openpyxl==3.0.3
openpyxl-image-loader==1.0.5
oss2==2.6.1
pbr==5.5.1
Pillow==6.2.1
powerline-status==2.7
pycryptodome==3.8.1
pyDes==2.0.1
PyMySQL==0.9.3
PyPDF2==1.26.0
python-consul==1.1.0
python-dateutil==2.8.0
pytz==2019.1
qiniu==7.2.6
redis==3.2.1
reportlab==3.5.42
requests==2.22.0
simplejson==3.16.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
stevedore==3.2.2
uritemplate==3.0.0
urllib3==1.25.2
vine==1.3.0
virtualenv==20.1.0
virtualenv-clone==0.5.4
Werkzeug==1.0.1
xmljson==0.2.0
zipp==3.4.0
